requirements.txt
python-dotenv
requests
fastapi